Add Collections.Set.Ordered, Dictionary.Ordered, and Binary.Collection types
Implements ordered collections with nested accessor pattern:

Collections.Set.Ordered:
- Hash table + array hybrid for O(1) operations
- Insertion order preserved
- Nested algebra accessor: set.algebra.union(), .intersection(),
  .subtract(), .symmetric.difference()

Collections.Dictionary.Ordered:
- Built on Set.Ordered for keys
- Nested accessors: dict.values.set(), .remove(), .modify()
  and dict.keys.index()
- Policy-based merge: dict.merge.keep.first/last()
- Typed error with Duplicate payload

Binary.Collection.Set (BitSet):
- Space-efficient set of non-negative integers
- One bit per element, word-level operations
- Nested algebra accessor matching Set.Ordered pattern

Binary.Collection.Array (BitArray):
- Dense boolean array, one bit per element
- RandomAccessCollection conformance
- O(1) append, subscript, toggle operations

All types follow PATTERN REQUIREMENTS.md:
- Nested accessor pattern as primary API
- Per-type typed errors with data-only payloads
- Conditional Sendable conformance
- CoW semantics where applicable

Tests: 91 new tests (Set.Ordered: 22, Dictionary.Ordered: 22,
BitSet: 26, BitArray: 21)

+++ Sources/Binary/Binary.Format.Bytes.Unit.swift
@@ -9,7 +9,7 @@ extension Binary.Format.Bytes {
     /// for both naming conventions.
     public struct Unit: Sendable, Equatable {
         /// The multiplier for this unit relative to bytes.
-        public let multiplier: Int
+        public let multiplier: Int64
 
         /// Symbol for decimal (SI) system (e.g., "KB").
         public let decimalSymbol: String
@@ -18,7 +18,7 @@ extension Binary.Format.Bytes {
         public let binarySymbol: String
 
         @usableFromInline
-        init(multiplier: Int, decimalSymbol: String, binarySymbol: String) {
+        init(multiplier: Int64, decimalSymbol: String, binarySymbol: String) {
             self.multiplier = multiplier
             self.decimalSymbol = decimalSymbol
             self.binarySymbol = binarySymbol
